# Implement lazy retrieval of series from object store.