--- a/.gitignore
+++ b/.gitignore
@@ -17,6 +17,9 @@ kube/.minikube
 data/
 test/e2e/e2e_*
 
+# Ignore benchmarks dir
+benchmarks/
+
 # Ignore promu artifacts.
 /.build
 /.release