--- a/cmd/thanos/store.go
+++ b/cmd/thanos/store.go
@@ -6,6 +6,7 @@ package main
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"time"
 
 	"github.com/alecthomas/units"
@@ -130,8 +131,8 @@ func (sc *storeConfig) registerFlag(cmd extkingpin.FlagClause) {
 	cmd.Flag("block-meta-fetch-concurrency", "Number of goroutines to use when fetching block metadata from object storage.").
 		Default("32").IntVar(&sc.blockMetaFetchConcurrency)
 
-	cmd.Flag("debug.series-batch-size", "The batch size when fetching series from object storage.").
-		Hidden().Default("10000").IntVar(&sc.seriesBatchSize)
+	cmd.Flag("debug.series-batch-size", "The batch size when fetching series from TSDB blocks. Setting the number too high can lead to slower retrieval, while setting it too low can lead to throttling caused by too many calls made to object storage.").
+		Hidden().Default(strconv.Itoa(store.SeriesBatchSize)).IntVar(&sc.seriesBatchSize)
 
 	sc.filterConf = &store.FilterConfig{}