--- a/cmd/thanos/store.go
+++ b/cmd/thanos/store.go
@@ -56,6 +56,7 @@ type storeConfig struct {
 	httpConfig                  httpConfig
 	indexCacheSizeBytes         units.Base2Bytes
 	chunkPoolSize               units.Base2Bytes
+	seriesBatchSize             int
 	maxSampleCount              uint64
 	maxTouchedSeriesCount       uint64
 	maxDownloadedBytes          units.Base2Bytes
@@ -128,6 +129,9 @@ func (sc *storeConfig) registerFlag(cmd extkingpin.FlagClause) {
 
 	cmd.Flag("block-meta-fetch-concurrency", "Number of goroutines to use when fetching block metadata from object storage.").
 		Default("32").IntVar(&sc.blockMetaFetchConcurrency)
+
+	cmd.Flag("debug.series-batch-size", "The batch size when fetching series from object storage.").
+		Hidden().Default("10000").IntVar(&sc.seriesBatchSize)
 
 	sc.filterConf = &store.FilterConfig{}
 
@@ -340,6 +344,7 @@ func runStore(
 		store.WithChunkPool(chunkPool),
 		store.WithFilterConfig(conf.filterConf),
 		store.WithChunkHashCalculation(true),
+		store.WithSeriesBatchSize(conf.seriesBatchSize),
 	}
 
 	if conf.debugLogging {