… pkg/store/bucket_test.go
… pkg/store/storepb/custom.go