The bucket store fetches series in a single blocking operation from object storage. This is likely not an ideal strategy when it comes to latency and resource usage. In addition, it causes the store to buffer everything in memory before starting to send results to queriers.

This commit modifies the series retrieval to use the proxy response heap and take advantage of the k-way merge used in the proxy store.

cc @GiedriusS @bwplotka 

Latest benchmarks
```
name                                                          old time/op    new time/op    delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-8             57.4ms ±14%    38.1ms ± 1%  -33.58%  (p=0.029 n=4+4)
BucketSeries/1000000SeriesWith1Samples/10of1000000-8            53.7ms ± 1%    38.2ms ± 1%  -28.85%  (p=0.029 n=4+4)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-8        521ms ± 3%     431ms ± 1%  -17.17%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/1of10000000-8           3.80ms ± 1%    4.17ms ± 1%   +9.85%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/100of10000000-8         3.79ms ± 0%    4.19ms ± 3%  +10.66%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-8    49.4ms ± 1%    44.3ms ± 1%  -10.45%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/1of10000000-8            135µs ± 1%     137µs ± 0%   +1.71%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/100of10000000-8          134µs ± 1%     137µs ± 1%   +2.05%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-8    12.1ms ± 0%    12.8ms ± 1%   +5.39%  (p=0.029 n=4+4)

name                                                          old alloc/op   new alloc/op   delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-8             60.9MB ± 0%    54.3MB ± 0%  -10.82%  (p=0.029 n=4+4)
BucketSeries/1000000SeriesWith1Samples/10of1000000-8            60.9MB ± 0%    54.3MB ± 0%  -10.82%  (p=0.029 n=4+4)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-8       1.29GB ± 0%    0.80GB ± 0%  -38.11%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/1of10000000-8           4.58MB ± 0%    6.06MB ± 0%  +32.40%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/100of10000000-8         4.58MB ± 0%    6.06MB ± 0%  +32.39%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-8     120MB ± 3%      86MB ± 1%  -28.19%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/1of10000000-8            200kB ± 0%     202kB ± 0%   +0.81%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/100of10000000-8          200kB ± 0%     202kB ± 0%   +0.87%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-8    46.8MB ± 0%    46.8MB ± 0%     ~     (p=0.057 n=4+4)

name                                                          old allocs/op  new allocs/op  delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-8              9.61k ± 0%     4.65k ± 0%  -51.61%  (p=0.029 n=4+4)
BucketSeries/1000000SeriesWith1Samples/10of1000000-8             9.71k ± 0%     4.76k ± 0%  -51.03%  (p=0.029 n=4+4)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-8        10.0M ± 0%     11.0M ± 0%   +9.79%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/1of10000000-8            1.08k ± 0%     0.71k ± 0%  -34.19%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/100of10000000-8          1.11k ± 0%     0.75k ± 0%  -32.88%  (p=0.029 n=4+4)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-8     1.00M ± 0%     1.10M ± 0%   +9.80%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/1of10000000-8              282 ± 0%       293 ± 0%   +3.90%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/100of10000000-8            282 ± 0%       293 ± 0%   +3.90%  (p=0.029 n=4+4)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-8      168k ± 0%      168k ± 0%   +0.03%  (p=0.029 n=4+4)

```

<!--
    Keep PR title verbose enough and add prefix telling
    about what components it touches e.g "query:" or ".*:"
-->

<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [ ] I added CHANGELOG entry for this change.
* [x] Change is not relevant to the end user.

## Changes

<!-- Enumerate changes you made -->

## Verification

<!-- How you tested it? How do you know it works? -->