@fpetkovski 2022-10-29
Some benchmarks with batched retrieval
```
name                                                          old time/op    new time/op    delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-8             54.5ms ± 2%    54.5ms ± 2%     ~     (p=1.000 n=5+5)
BucketSeries/1000000SeriesWith1Samples/10of1000000-8            54.3ms ± 2%    54.6ms ± 2%     ~     (p=0.548 n=5+5)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-8        510ms ± 1%     431ms ± 2%  -15.45%  (p=0.008 n=5+5)
BucketSeries/100000SeriesWith100Samples/1of10000000-8           3.91ms ± 0%    3.86ms ± 1%   -1.28%  (p=0.008 n=5+5)
BucketSeries/100000SeriesWith100Samples/100of10000000-8         3.90ms ± 1%    3.85ms ± 1%   -1.28%  (p=0.032 n=5+5)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-8    50.3ms ± 2%    41.6ms ± 1%  -17.29%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/1of10000000-8            134µs ± 1%     154µs ± 2%  +14.26%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/100of10000000-8          135µs ± 1%     151µs ± 1%  +11.27%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-8    12.2ms ± 1%    12.2ms ± 1%     ~     (p=0.421 n=5+5)

name                                                          old alloc/op   new alloc/op   delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-8             60.9MB ± 0%    61.0MB ± 0%   +0.23%  (p=0.008 n=5+5)
BucketSeries/1000000SeriesWith1Samples/10of1000000-8            60.9MB ± 0%    61.0MB ± 0%   +0.24%  (p=0.008 n=5+5)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-8       1.29GB ± 0%    0.89GB ± 0%  -30.90%  (p=0.016 n=4+5)
BucketSeries/100000SeriesWith100Samples/1of10000000-8           4.58MB ± 0%    4.67MB ± 0%   +1.95%  (p=0.008 n=5+5)
BucketSeries/100000SeriesWith100Samples/100of10000000-8         4.58MB ± 0%    4.67MB ± 0%   +1.94%  (p=0.008 n=5+5)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-8     120MB ± 2%      84MB ± 1%  -30.20%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/1of10000000-8            200kB ± 0%     277kB ± 0%  +38.16%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/100of10000000-8          200kB ± 0%     276kB ± 0%  +38.01%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-8    46.8MB ± 0%    47.1MB ± 0%   +0.62%  (p=0.008 n=5+5)

name                                                          old allocs/op  new allocs/op  delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-8              9.63k ± 0%    11.36k ± 0%  +17.99%  (p=0.016 n=4+5)
BucketSeries/1000000SeriesWith1Samples/10of1000000-8             9.71k ± 0%    11.49k ± 0%  +18.35%  (p=0.008 n=5+5)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-8        10.0M ± 0%     12.1M ± 0%  +20.20%  (p=0.008 n=5+5)
BucketSeries/100000SeriesWith100Samples/1of10000000-8            1.08k ± 0%     1.25k ± 0%  +15.90%  (p=0.008 n=5+5)
BucketSeries/100000SeriesWith100Samples/100of10000000-8          1.11k ± 0%     1.29k ± 0%  +16.08%  (p=0.008 n=5+5)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-8     1.00M ± 0%     1.21M ± 0%  +20.15%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/1of10000000-8              282 ± 0%       286 ± 0%   +1.42%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/100of10000000-8            282 ± 0%       286 ± 0%   +1.42%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-8      168k ± 0%      168k ± 0%   +0.02%  (p=0.008 n=5+5)
```