@fpetkovski 2022-10-31
This should now be ready for review. I've added the latest benchmark results to the description of the PR.

I don't think the benchmark is realistic though, I feel it overestimates the performance of the current implementation, and underestimates the retrieval performance. The first reason for this is because the benchmark uses an in-memory dataset so chunk retrieval is instantaneous. The second one is because it buffers all received series, so their memory cannot be released when they are sent over a gRPC channel. I would expect performance in a realistic scenario to be better than what is predicted.