@GiedriusS 2022-11-07
Have you tried this with "real" data? Is there any noticeable improvement? :thinking: