@fpetkovski 2022-11-09
@GiedriusS In a staging environment things look ~20-30% faster for a 30d range query, so from 14s to 10s. I would like to double check memory usage this week in a prod-like environment to make sure there are no leaks or regressions.