@fpetkovski 2022-11-10
I was testing this today and so far it looks good. I haven't seen any issues with memory usage or leaks.