@fpetkovski 2022-11-14
I noticed that expanding postings and reading the index is also a non-trivial amount of work, so maybe that is also a significant factor.